processing_service/requirements.txt
confluent-kafka>=2.0.0
python-dotenv>=1.0.0